+++ test/cookbooks/docker_test/recipes/image_lwrp.rb
@@ -17,8 +17,6 @@
 # limitations under the License.
 #
 
-include_recipe "docker"
-
 docker_image "base"
 docker_image "busybox"